backend/requirements.txt
# --- Web Server Dependencies ---
# Using a known stable version set
uvicorn[standard]==0.29.0
fastapi==0.110.0

# --- Real-time Communication (FIXED for compatibility with client v4) ---
python-engineio==4.8.0
python-socketio==5.10.0

# --- Core Simulation Engine Dependencies ---
numpy
pandas
scipy
PyYAML